Renamed ColorInput -> ColorPicker
