+++ src/components/inputs/ColorPicker.vue
