+++ src/components/items/ItemEditorPanel.vue
@@ -8,7 +8,7 @@ import { makeDateRange, type DateRange } from '@/utils/dates'
 import { useInterfaceStore } from '@/stores/interface'
 import { endDateAfterStart, rangeDatesValid, required } from '@/utils/validation'
 import DateRangePicker from '../DateRangePicker.vue'
-import ColorInput from '../inputs/ColorInput.vue'
+import ColorInput from '../inputs/ColorPicker.vue'
 
 // External State
 const model = defineModel<boolean>()
